feat: merge all fetch functions into fetch.py
